requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiosmtpd==1.4.4.post2
    # via -r requirements.in
atpublic==4.0
    # via aiosmtpd
attrs==23.1.0
    # via aiosmtpd
blinker==1.7.0
    # via flask
bootstrap-flask==2.3.2
    # via -r requirements.in
click==8.1.7
    # via
    #   flask
    #   rq
dnspython==2.4.2
    # via email-validator
email-validator==2.1.0.post1
    # via -r requirements.in
flask==3.0.0
    # via
    #   -r requirements.in
    #   bootstrap-flask
    #   flask-httpauth
    #   flask-login
    #   flask-mailman
    #   flask-moment
    #   flask-sqlalchemy
    #   flask-wtf
flask-httpauth==4.8.0
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-mailman==1.0.0
    # via -r requirements.in
flask-moment==1.0.5
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
flask-wtf==1.2.1
    # via -r requirements.in
greenlet==3.0.1
    # via sqlalchemy
idna==3.4
    # via email-validator
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
packaging==23.2
    # via flask-moment
pyjwt==2.8.0
    # via -r requirements.in
redis==5.0.1
    # via rq
rq==1.15.1
    # via -r requirements.in
sqlalchemy==2.0.23
    # via flask-sqlalchemy
typing-extensions==4.8.0
    # via sqlalchemy
werkzeug==3.0.1
    # via
    #   flask
    #   flask-login
wtforms==3.1.1
    # via
    #   bootstrap-flask
    #   flask-wtf
